requirements.txt
requests==2.31.0
python-chess==1.999
openai>=1.6.0
python-dotenv==1.0.0 